requirements.txt
h5py==2.9.0
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
numpy==1.16.1
opencv-contrib-python==4.0.0.21
opencv-python==4.0.0.21
pandas==0.24.0
Pillow==5.4.1
protobuf==3.6.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
scipy==1.2.0
selenium==3.141.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
